Allow num0 to be any decimal number

+++ calc.py
@@ -1,12 +1,13 @@
 def compute(expression):
-    values = expression.split(' ')
-    num0 = int(values[0])
+    values = expression.split(" ")
+    num0 = float(values[0])
     operator = values[1]
     num1 = int(values[2])
-    if operator == '+':
+    if operator == "+":
         return num0 + num1
-    elif operator == '-':
+    elif operator == "-":
         return num0 - num1
     else:
-        print('unknown operator!')
+        print("unknown operator!")
         return None
+
